refactor: apply Vercel React Best Practices
- Extracted global `keydown` event listener logic into a deduplicated `useGlobalKeydown` custom hook based on the `client-event-listeners.md` best practice.
- Applied `useGlobalKeydown` hook to `app/prs/6th-anniversary-revenue-report/ImageModal.tsx`, `app/games/typing/page.tsx`, and `app/games/mochinoa-jump/page.tsx`, replacing multiple `window.addEventListener` instances with a single global listener.
- Fixed typo `"ues client"` to `"use client"` in `app/guess-number/GuessNumberDialog.tsx`.
