+++ app/prs/6th-anniversary-revenue-report/ImageModal.tsx
@@ -5,6 +5,7 @@ import { AnimatePresence, motion } from "motion/react";
 import type { StaticImageData } from "next/image";
 import Image from "next/image";
 import { useCallback, useEffect, useEffectEvent, useState } from "react";
+import { useGlobalKeydown } from "../../../components/hooks/useGlobalKeydown";
 import styles from "./page.module.css";
 
 interface ImageModalProps {
@@ -17,20 +18,6 @@ const zoomInIconLarge = <ZoomIn size={32} />;
 const zoomInIconSmall = <ZoomIn size={20} />;
 const zoomOutIconSmall = <ZoomOut size={20} />;
 const closeIcon = <X size={24} />;
-
-// BEST PRACTICE: Deduplicate Global Event Listeners (client-event-listeners.md)
-const keydownCallbacks = new Set<(e: KeyboardEvent) => void>();
-let isGlobalKeydownListenerAdded = false;
-
-function addGlobalKeydownListener() {
-	if (typeof window === "undefined" || isGlobalKeydownListenerAdded) return;
-	window.addEventListener("keydown", (e: KeyboardEvent) => {
-		for (const cb of keydownCallbacks) {
-			cb(e);
-		}
-	});
-	isGlobalKeydownListenerAdded = true;
-}
 
 export function ImageModal({ src, alt }: ImageModalProps) {
 	const [isOpen, setIsOpen] = useState(false);
@@ -51,11 +38,10 @@ export function ImageModal({ src, alt }: ImageModalProps) {
 		if (e.key === "Escape") closeModal();
 	});
 
+	useGlobalKeydown(onEscEvent);
+
 	useEffect(() => {
-		addGlobalKeydownListener();
-		keydownCallbacks.add(onEscEvent);
 		return () => {
-			keydownCallbacks.delete(onEscEvent);
 			document.body.style.overflow = "unset";
 		};
 	}, []);
